% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saymyname.R
\name{saymyname}
\alias{saymyname}
\title{Saying back someone's name}
\usage{
saymyname(test_name, my_name = "Cavin")
}
\arguments{
\item{test_name}{Name to be tested}

\item{my_name}{Name given to the function}
}
\value{
What? for non-names, Who? for wrong names, and statment for correct names
}
\description{
A few paragraphs of description if needed
}
\details{
None needed
}

